Create a Default Image System

+++ src/main/java/me/noxiuam/noxlib/command/ticket/CloseTicket.java
@@ -4,14 +4,23 @@ import me.noxiuam.noxlib.NoxLib;
 import me.noxiuam.noxlib.command.Command;
 import me.noxiuam.noxlib.command.CommandContext;
 
-public class CloseTicket extends Command {
+public class CloseTicket extends Command
+{
 
-    public CloseTicket() {
+    public CloseTicket()
+    {
         super("close", "Closes a ticket.", "$close");
     }
 
     @Override
-    public void execute(CommandContext ctx) {
+    public void execute(CommandContext ctx)
+    {
+        if (!ctx.getChannel().getName().startsWith("ticket-"))
+        {
+            ctx.getChannel().sendMessage(NoxLib.getInstance().getMessageUtil().createEmbedWithThumbnail("Error closing ticket", "This channel is not a ticket!", NoxLib.getInstance().getDefaultImage()).build()).queue();
+            return;
+        }
+
         ctx.getChannel().delete().queue();
         NoxLib.getInstance().getOpenTickets().remove(ctx.getAuthor().getAsTag());
     }
